refactor(agent): widen @agent/index door, retire registry/dir leaf imports
Second consolidation step from the -08-21 readiness pass (§7b). Hosts were
deep-reaching @agent/index/{agentRegistry,platformAgentDirectories,
BundledAgentDirectories} for symbols the @agent/index barrel already exports;
only createPlatformAgentDirectories and BUNDLED_AGENT_DIRECTORY_NAMES were
genuinely absent (both agent-registry/directory surface). Add those two
re-exports and route the six leaf importers (cli initPlatform; desktop
main/index, platform/index, platform/paths, settingsController; extension
AgentDirectoryManager) through @agent/index.

Baseline: cli 9->8, desktop 9->6, extension 11->10. Combined with the prior
barrel commit, net -8 host deep-import specifiers this session; four of eight
planned Tier-1 doors now in place. The @agent/core/state/executionRequests leaf
is left as-is: a desktop dynamic import() (ratchet-counted) blocks a clean door
move without eager-loading it. Behavior-preserving; no runtime logic moved.

typecheck clean; deep-import ratchet green; dead-code ratchet no new findings;
eslint + prettier clean; architecture suite 104/104.
